requirements.txt
acres==0.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
cymem==2.0.11
DateTime==5.5
distro==1.9.0
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
etelemetry==0.3.1
exceptiongroup==1.2.2
faiss-cpu==1.11.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fpdf==1.7.2
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
narwhals==1.40.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.8.1
numpy==1.26.4
ollama==0.5.1
openai==1.75.0
openapi==2.0.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
prov==2.0.1
puremagic==1.29
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydot==4.0.0
Pygments==2.19.1
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==2.6.1
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.15.2
toml==0.10.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.5
tqdm==4.67.1
traits==7.0.2
transformers==4.37.2
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
zope.interface==7.2
